reorganize read ahead tasks
